requirements.txt
pandas
discord.py
python-dotenv
rich
